simplify build scripts
