+++ platform/core/.webpack/webpack.dev.js
