+++ platform/viewer/.webpack/webpack.commonjs.js
@@ -1,9 +1,9 @@
 const path = require('path');
 const merge = require('webpack-merge');
-const webpack = require('webpack');
-const webpackCommon = require('./../../../.webpack/webpack.common.js');
+const webpackCommon = require('./../../../.webpack/webpack.commonjs.js');
 //
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const fontsToJavaScriptRule = require('./rules/fontsToJavaScript.js');
 // const
 const SRC_DIR = path.join(__dirname, '../src');
 const DIST_DIR = path.join(__dirname, '../dist');
@@ -12,8 +12,6 @@ module.exports = (env, argv) => {
   const commonConfig = webpackCommon(env, argv, { SRC_DIR, DIST_DIR });
 
   return merge(commonConfig, {
-    // https://webpack.js.org/configuration/mode/#mode-production
-    mode: 'production',
     entry: {
       bundle: `${SRC_DIR}/index-umd.js`,
     },
@@ -39,27 +37,8 @@ module.exports = (env, argv) => {
       libraryTarget: 'umd',
       filename: 'index.umd.js',
     },
-    /**
-     * For CommonJS, we want to bundle whatever font we've landed on. This allows
-     * us to reduce the number of script-tags we need to specify for simple use.
-     *
-     * PWA will grab these externally to reduce bundle size (think code split),
-     * and cache the grab using service-worker.
-     */
     module: {
-      rules: [
-        {
-          test: /\.(ttf|eot|woff|woff2)$/i,
-          use: [
-            {
-              loader: 'file-loader',
-              options: {
-                name: '[name].[ext]',
-              },
-            },
-          ],
-        },
-      ],
+      rules: [fontsToJavaScriptRule],
     },
     plugins: [
       // Clean output.path
