+++ postcss.config.js
@@ -1,9 +1,13 @@
-module.exports = {
-  // parser: "sugarss",
-  parser: false,
-  plugins: {
-    "postcss-import": {},
-    "postcss-preset-env": {},
-    cssnano: {}
-  }
+module.exports = function(ctx) {
+  ctx = ctx || {};
+  ctx.env = ctx.env || 'development';
+
+  return {
+    map: ctx.env === 'development' ? ctx.map : false,
+    plugins: {
+      'postcss-import': {},
+      'postcss-preset-env': {},
+      cssnano: ctx.env === 'production' ? {} : false,
+    },
+  };
 };
